+++ e2e/tests/constants.js
@@ -11,6 +11,7 @@ export const fixtures = [
   'html',
   'gallery-layouts',
   'old-image-format',
+  'hashtag-and-link',
 ];
 
 export const fixturesToTestOnSeo = ['images'];
@@ -20,7 +21,4 @@ export const DEFAULT_DESKTOP_BROWSERS = [
   { width: 1440, height: 900, name: 'firefox' },
 ];
 
-export const DEFAULT_MOBILE_BROWSERS = [
-  { deviceName: 'iPhone X' },
-  { deviceName: 'iPad' },
-];
+export const DEFAULT_MOBILE_BROWSERS = [{ deviceName: 'iPhone X' }, { deviceName: 'iPad' }];